Refactor project structure for IBM DFDL testing integration:
- Migrated IBM DFDL logic to `processor` and `plugin` modules.
- Added custom `IBMDFDLPlugin` to simplify test process
- Moved IBM DFDL files to be copied in to `ibm-dfdl` directory.
- Updated `.gitignore` to reflect new file structure and streamlined organization.
- Added some missing imports in `IBM_DFDL.scala`.
- Enhanced `build.sbt` with better support for IBM DFDL testing.
- Added `sbt-assembly` plugin for packaging fat jars
- Updated validation logic in `IBM_DFDL.scala`.
- Simplified setup steps in `README.md`.
- Removed obsolete plugin and updated `sbt.version`.
- Added `commonSettings` for shared SBT configurations across modules.
