+++ .gitignore
@@ -4,37 +4,5 @@
 #
 # Similarly, we're not saving the sample files from the IBM DFDL samples.
 #
-lib/*.jar
-src/test/resources/company.xsd
-src/test/resources/company.xml
-src/test/resources/company.txt
-src/test/resources/IBMdefined/RecordSeparatedFieldFormat.xsd
-*.sbt
-!build.sbt
-.cache-main
-.cache-tests
-target
+ibm-dfdl/
 target/
-lib_managed
-src_managed
-resource_managed
-.history
-.cache
-.idea
-.idea_modules
-project/errors
-.~*#
-.scala_dependencies
-.settings
-# ignore all emacs backup files
-*~
-#
-# FOR EMACS ENSIME
-#
-.ensime_cache
-# these are autosave emacs files
-\#*#
-.\#*
-.\#.*
-/out/
-/bin/
