+++ README.md
@@ -39,20 +39,20 @@ In the IBM ACE product you will find a DFDL jar that you can unzip--various
 files must be copied out of this jar to enable building and testing the cross
 tester.
 
-1. Copy the IBM DFDL jars into the ``lib`` subdirectory of this project.
+1. Copy the IBM DFDL jars into the ``ibm-dfdl/lib`` subdirectory of this project.
    
    Note that you must also copy the ``samples/dfdlsample_java.jar`` into the
    ``lib`` subdirectory, as that also has class files in it that this test rig
-   uses.
+   uses for compilation.
 
    ```
-   cp ~/path/to/IBMDFDL/*.jar lib/
-   cp ~/path/to/IBMDFDL/samples/*.jar lib/
+   cp ~/path/to/IBMDFDL/*.jar ibm-dfdl/lib/
+   cp ~/path/to/IBMDFDL/samples/*.jar ibm-dfdl/lib/
    ```
    The resulting directory looks like this.
 
    ```
-   lib
+   ibm-dfdl/lib
    ├── dfdlsample_java.jar
    ├── emf.common_2.6.0.jar
    ├── emf.ecore_2.6.1.jar
@@ -65,46 +65,42 @@ tester.
    └── xsd_2.6.0.jar
    ```
 
-2. Copy the sample "company" files to ``src/test/resources``
+2. Copy the sample "company" files to ``ibm-dfdl/src/test/resources``
 
    ```
-   cp ~/path/to/IBMDFDL/company.* src/test/resources
+   cp ~/path/to/IBMDFDL/company.* ibm-dfdl/src/test/resources
    ```
 
-3. Copy the sample schema file into ``src/test/resources/IBMdefined/``
+3. Copy the sample schema file into ``ibm-dfdl/src/test/resources/IBMdefined/``
 
    ```
-   cp ~/path/to/IBMDFDL/IBMdefined/RecordSeparatedFieldFormat.xsd  src/test/resources/IBMdefined
+   cp ~/path/to/IBMDFDL/IBMdefined/RecordSeparatedFieldFormat.xsd  ibm-dfdl/src/test/resources/IBMdefined
    ```
 
 The files in step 2 and 3 are examples created by IBM which this test rig will
 invoke using TDML to show that everything is working properly.
 
-After completing those steps, the resulting tree under ``src/test/resources``
+After completing those steps, the resulting tree under ``ibm-dfdl/src/test/resources``
 should look like:
 
-```
-src/test/resources/
-├── companySelfContained.tdml       - supplied by this cross test rig.
-├── company.tdml                    - supplied by this cross test rig.
-├── company.txt (copy from IBM DFDL)
-├── company.xml (copy from IBM DFDL)
-├── company.xsd (copy from IBM DFDL)
-├── crossTestRigTestSchema.dfdl.xsd - supplied by this cross test rig, to test the rig itself.
-├── crossTestRigTests.tdml          - supplied by this cross test rig, to test the rig itself.
-└── IBMdefined
-    └── RecordSeparatedFieldFormat.xsd (copy from IBM DFDL)
-```
+   ```
+   ibm-dfdl/src/test/resources/
+   ├── company.txt (copy from IBM DFDL)
+   ├── company.xml (copy from IBM DFDL)
+   ├── company.xsd (copy from IBM DFDL)
+   └── IBMdefined
+       └── RecordSeparatedFieldFormat.xsd (copy from IBM DFDL)
+   ```
 
 ## Build & Test
 
 To build and test the cross tester to make sure all dependencies and samples
 files have been copied correctly, run the following commands:
 
-```
-sbt compile
-sbt test
-```
+   ```
+   sbt compile
+   sbt test
+   ```
 
 This runs the cross testers own self-tests and two versions of the IBM-supplied
 example tests. One version is with a TDML file, ``company.tdml``, which refers
@@ -118,59 +114,37 @@ non-portability issue.
 
 ## Testing Other Test Suites
 
+> !IMPORTANT!
+> The below steps create a jar with the IBM DFDL jars embedded so has the
+> same distribution restrictions as those jars
+
 To use this cross-tester and run other test suites such as those in daffodil
 or those for the DFDLSchemas projects on github, follow these steps:
 
 1. Publish the IBM DFDL Cross Tester jar so that other software modules can find
-   it, with the following command:
+it, with the following command:
 
    ```
-   sbt publishLocal
+      sbt publishLocal
    ```
 
-2. Setup the sbt plugin that makes it easy to use this cross test rig, with the
-   following command:
-
+2. Add the following line to your ``project/plugins.sbt`` file to make the IBM
+DFDL Cross Tester plugin available to your schema project:
+   
    ```
-   mkdir -p ~/.sbt/1.0/plugins/
-   cp plugin/ibmDFDLCrossTesterSBTPlugin.scala ~/.sbt/1.0./plugins/
-   ```
-
-3. Modify ``~/.sbt/1.0./plugins/bmDFDLCrossTesterSBTPlugin.scala`` to include
-   the path to the lib directory containing the IBM DFDL jars, for example:
-
-   ```scala
-   val ibmDFDLJarDirectory = file("/home/user/git/ibmDFDLCrossTester/lib/")
+      addSbtPlugin("org.apache.daffodil.tdml.processor" % "sbt-ibm-dfdl" % "1.0.0-SNAPSHOT")
    ```
 
-The above steps allow one to easily run a DFDL schema project against Daffodil
-or IBM by changing only one line in the ``build.sbt`` file.
-
-If your build.sbt uses old-style SBT configuration, add this line anywhere in
-your file to enable the IBM cross tester:
-
-```
-IBMDFDLCrossTesterPlugin.settings
-```
-
-If you use the newer SBT config file style, add the settings like your
-``project`` like so:
-
-```
-    .settings(IBMDFDLCrossTesterPlugin.settings)
-```
-
-With the settings added above, he plugin will modify the classpath so that the
-IBM DFDL Cross Tester's ibm-tdml-processor will be used instead of the
-daffodil-tdml-processor. Then when you run ``sbt test`` in your schema project,
-it will use IBM DFDL to run the tests.
+With the settings added above, the cross tester will use IBM DFDL to run the tests 
+when you run ``sbt ibmTest`` in your schema project. To run the tests without
+using IBM DFDL, use ``sbt test`` instead.
 
 Keep in mind the tests must have "ibm" as a member of a TDML testSuite
 ``defaultImplementations`` attribute, or a TDML parserTestCase or
 unparserTestCase ``implementations`` attribute. Otherwise the test is skipped
 for that implementation. Tests intended to be portable should list
 ``defaultImplementations="ibm daffodil"`` (or on the test case,
-``implemenations="ibm daffodil``) so that both implementations will attempt
+``implementations="ibm daffodil``) so that both implementations will attempt
 to run the test. A test that works only on one of the implementations
 should leave out the other implementation.
