+++ processor/src/test/resources/companySelfContained.tdml
