+++ project/plugins.sbt
@@ -1,1 +1,2 @@
 addSbtPlugin("org.apache.daffodil" % "sbt-daffodil" % "1.+")
+addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "2.3.1")
